antlr-metamodel-parser/requirements.txt
click==8.1.3
Flask==2.3.0
flask[async]==2.3.0
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.3.0
requests==2.31.0
langchain==0.0.321
python-dotenv==1.0.0
flask_cors==4.0.0
llama_index==0.8.50
openai==0.28.1
llama_hub==0.0.39
asyncio== 3.4.3
google-search-results==2.4.2